+++ components/JustMap.js
@@ -1,8 +1,3 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
 
 import React, { Component } from 'react';
 import {
@@ -20,7 +15,6 @@ import ListItem from './ListItem';
 import * as firebase from 'firebase';
 import Firebase from "../includes/firebase";
 
-
 const { width, height } = Dimensions.get('window');
 
 const ASPECT_RATIO = width / height;
@@ -28,20 +22,11 @@ const LATITUDE = 37.78825;
 const LONGITUDE = -122.4324;
 const LATITUDE_DELTA = 0.0922;
 const LONGITUDE_DELTA = LATITUDE_DELTA * ASPECT_RATIO;
-let id = 0;
-
 
 const SideMenu = require('react-native-side-menu');
 const Menu = require('./Menu');
 
-
-
 export default class JustMap extends React.Component {
-
-
-
-
-
 
   constructor(props) {
     super(props);
@@ -52,9 +37,9 @@ export default class JustMap extends React.Component {
         latitudeDelta: LATITUDE_DELTA,
         longitudeDelta: LONGITUDE_DELTA,
       },
-      markers: [],
       initialPosition: 'unknown',
       polylines: [],
+      locations: [],
       dataSource: new ListView.DataSource({
         rowHasChanged: (row1, row2) => row1 !== row2,
       })
@@ -64,7 +49,6 @@ export default class JustMap extends React.Component {
 
     // firebase reference
     this.itemsRef = this.getRef().child('locations');
-
   }
 
     // firebase Example
@@ -78,46 +62,53 @@ export default class JustMap extends React.Component {
          snap.forEach((child) => {
            items.push({
              title: child.val().title,
-             _key: child.key
+             coordinate: child.val().coordinates,
+             key: child.getKey(),
+             name: child.val().name,
+             title: child.val().title,
+             description: child.val().description,
+             image: child.val().image,
+             imagePin: child.val().imagePin,
+             datePin:  child.val().datePin
            });
          });
 
          this.setState({
-           dataSource: this.state.dataSource.cloneWithRows(items)
+           dataSource: this.state.dataSource.cloneWithRows(items),
+           locations: items
          });
+
+         console.log("items", items);
      });
     }
 
    componentDidMount() {
      this.listenForItems(this.itemsRef);
-
-
-    //  navigator.geolocation.getCurrentPosition(
-    //    (position) => {
-    //      var initialPosition = JSON.stringify(position);
-    //      this.setState({initialPosition});
-    //    },
-    //    (error) => alert(JSON.stringify(error)),
-    //    {enableHighAccuracy: true, timeout: 20000, maximumAge: 1000}
-    //  );
-    //  Check here : https://facebook.github.io/react-native/docs/permissionsandroid.html
-
    }
 
-   _addItem() {
-      this.itemsRef.push({ title: "testtest" });
-   }
+   //  navigator.geolocation.getCurrentPosition(
+   //    (position) => {
+   //      var initialPosition = JSON.stringify(position);
+   //      this.setState({initialPosition});
+   //    },
+   //    (error) => alert(JSON.stringify(error)),
+   //    {enableHighAccuracy: true, timeout: 20000, maximumAge: 1000}
+   //  );
+   //  Check here : https://facebook.github.io/react-native/docs/permissionsandroid.html
+
 
    _addLocationToFirebase(title, coordinates) {
-      this.itemsRef.push({ title: title, coordinates: coordinates });
+      this.itemsRef.push({
+        title: title,
+        coordinates: coordinates,
+        name: 'New Pin',
+        title: 'title',
+        description: 'description',
+        image: flagBlackImg,
+        imagePin: 'https://media.licdn.com/mpr/mpr/shrinknp_400_400/p/3/005/01b/27a/240ddec.jpg',
+        datePin:  Date()
+       });
    }
-
-   _renderItem(item) {
-     return (
-       <ListItem item={item} />
-     );
-   }
-  // end firebase Example
 
     onDrageEndMarker(e) {
       this.setState({ x: e.nativeEvent.coordinate })
@@ -130,29 +121,13 @@ export default class JustMap extends React.Component {
     }
 
     onLongPressCreateMarker(e) {
-      this.setState({
-        markers: [
-          ...this.state.markers,
-          {
-            coordinate: e.nativeEvent.coordinate,
-            key: id++,
-            name: 'New Pin',
-            title: 'title',
-            description: 'description' + id,
-            image: flagBlackImg,
-            imagePin: 'https://media.licdn.com/mpr/mpr/shrinknp_400_400/p/3/005/01b/27a/240ddec.jpg',
-            datePin:  Date(),
-          },
-        ],
-      });
 
-      this._addLocationToFirebase('test_'+id, e.nativeEvent.coordinate);
+      this._addLocationToFirebase('test', e.nativeEvent.coordinate);
 
       const { editing } = this.state;
       if (!editing) {
         this.setState({
           editing: {
-            id: id++,
             coordinates: [e.nativeEvent.coordinate],
           },
         });
@@ -174,31 +149,30 @@ export default class JustMap extends React.Component {
     return (
       <View style={styles.container}>
 
-
             <MapView
               provider={this.props.provider}
               style={styles.map}
               initialRegion={this.state.region}
               onLongPress = {this.onLongPressCreateMarker}>
+              <MapView.Marker
+                title="This is a title"
+                description="This is a description"
+                coordinate={this.state.region}
+              />
 
-
-              {this.state.markers.map((marker,i) =>{
+              {this.state.locations.map((location,i) =>{
                 return (
                   <MapView.Marker
-                    key={i}
+                    key={location.key}
                     draggable
-                    {... marker}
-                    onDragEnd={this.onDrageEndMarker}
-                    onPress={this.onPressMarker}
+                    {... location}
                     >
                       <View style={styles.marker}>
-                        <Text style={styles.text}>{marker.name}</Text>
+                        <Text style={styles.text}>{location.name}</Text>
                       </View>
                   </MapView.Marker>
-
                 )
               })}
-
 
               {this.state.polylines.map(polyline => (
                 <MapView.Polyline
@@ -220,15 +194,6 @@ export default class JustMap extends React.Component {
               }
 
             </MapView>
-
-            <ListView
-              dataSource={this.state.dataSource}
-              renderRow={this._renderItem.bind(this)}
-              enableEmptySections={true}
-            />
-
-            <Button onPress={this._addItem.bind(this)} title="Add to FireBase test" />
-
       </View>
     );
   }
@@ -245,7 +210,6 @@ const styles = StyleSheet.create({
       justifyContent: 'flex-end',
       alignItems: 'center',
     },
-
     marker: {
       marginLeft: -18,
       marginTop: 0,
